scheduler/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-trade-api==3.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Testing==0.8.1
frozendict==2.4.2
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
msgpack==1.0.3
multidict==6.0.5
multitasking==0.0.11
numpy==1.24.4
openai==1.25.1
packaging==24.0
pandas==2.0.3
peewee==3.17.3
pluggy==1.5.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
yarl==1.9.4
yfinance==0.2.38
zipp==3.18.1
azure-cosmos==4.3.1
pytest
matplotlib
datetime
schedule
alpaca-py
joblib==1.3.2
Pillow
jsonify
pyparsing
cycler
kiwisolver
python-dateutil
pycaret[full]
ipython
traitlets
dask[dataframe]
